Vendor_Project_Base_Folder/requirements.txt
absl-py==2.1.0
annotated-types==0.6.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
Django==5.0.3
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.15.1
dm-tree==0.1.8
Flask==3.0.2
flatbuffers==24.3.7
fonttools==4.50.0
gast==0.5.4
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.2
google-generativeai==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
h5py==3.10.0
html5lib==1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
keras==3.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
MouseInfo==0.1.3
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.10.0
oscrypto==1.3.0
packaging==24.0
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
PyGetWindow==0.0.9
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
PyMsgBox==1.0.9
pyparsing==3.1.2
pypdf==4.2.0
pyperclip==1.8.2
pypng==0.20220715.0
PyRect==0.2.0
PyScreeze==0.1.30
python-bidi==0.4.2
python-dateutil==2.9.0.post0
pytweening==1.2.0
PyWavelets==1.5.0
pywhatkit==5.4
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
setuptools==69.2.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
threadpoolctl==3.3.0
tinycss2==1.2.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.1
wheel==0.43.0
wikipedia==1.4.0
wrapt==1.16.0
xhtml2pdf==0.2.15
